💎feat : List모델 생성 및 MenuType 이동
